Execute ApiRequest
