+++ src/main/java/org/jenkinsci/plugins/docker/swarm/DeadAgentServiceReaperActor.java
@@ -5,7 +5,6 @@ import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
 import akka.actor.Props;
 import jenkins.model.Jenkins;
-import org.jenkinsci.plugins.docker.swarm.docker.api.DockerApiRequest;
 import org.jenkinsci.plugins.docker.swarm.docker.api.response.ApiException;
 import org.jenkinsci.plugins.docker.swarm.docker.api.response.SerializationException;
 import org.jenkinsci.plugins.docker.swarm.docker.api.service.DeleteServiceRequest;
@@ -37,14 +36,14 @@ public class DeadAgentServiceReaperActor extends AbstractActor {
             final DockerSwarmPlugin swarmPlugin = Jenkins.getInstance().getPlugin(DockerSwarmPlugin.class);
             final ActorSystem as = swarmPlugin.getActorSystem();
             String dockerSwarmApiUrl = DockerSwarmCloud.get().getDockerSwarmApiUrl();
-            final Object result = new DockerApiRequest(new ListServicesRequest(dockerSwarmApiUrl,"label","ROLE=jenkins-agent")).execute();
+            final Object result = new ListServicesRequest(dockerSwarmApiUrl,"label","ROLE=jenkins-agent").execute();
                 for(ScheduledService service : (List<ScheduledService>)getResult(result,List.class) ) {
-                    Object tasks = new DockerApiRequest( new ListTasksRequest(dockerSwarmApiUrl, "service", service.Spec.Name)).execute();
+                    Object tasks =  new ListTasksRequest(dockerSwarmApiUrl, "service", service.Spec.Name).execute();
                     if(tasks != null) {
                         for(Task task : (List<Task>)getResult(tasks,List.class)){
                             if(task.isComplete()){
                                 LOGGER.info("Reaping service: "+service.Spec.Name );
-                                new DockerApiRequest(new DeleteServiceRequest(dockerSwarmApiUrl,service.Spec.Name)).execute();
+                                new DeleteServiceRequest(dockerSwarmApiUrl,service.Spec.Name).execute();
                                 break;
                             }
                         }
