+++ src/main/java/org/jenkinsci/plugins/docker/swarm/docker/api/error/ErrorMessage.java
@@ -1,0 +1,5 @@
+package org.jenkinsci.plugins.docker.swarm.docker.api.error;
+
+public class ErrorMessage{
+    public String message;
+}
